+++ src/general_mha_model.h
@@ -16,13 +16,13 @@ class GeneralMHAModel : public torch::nn::Module {
     public:
         GeneralMHAModel(
             const Shard& shard_,
-            int64_t vocab_size_,
-            int64_t embed_dim_,
-            int64_t hidden_dim_,
-            int64_t num_heads_,
-            int64_t num_kv_heads_,
-            int64_t head_dim_,
-            int64_t max_seq_len_,
+            int vocab_size_,
+            int embed_dim_,
+            int hidden_dim_,
+            int num_heads_,
+            int num_kv_heads_,
+            int head_dim_,
+            int max_seq_len_,
             float_t rope_scaling_,
             bool is_cache_enabled = false
         );
@@ -39,13 +39,13 @@ class GeneralMHAModel : public torch::nn::Module {
 
     private:
         const Shard& shard;
-        int64_t vocab_size;
-        int64_t embed_dim;
-        int64_t hidden_dim;
-        int64_t num_heads;
-        int64_t num_kv_heads;
-        int64_t head_dim;
-        int64_t max_seq_len;
+        int vocab_size;
+        int embed_dim;
+        int hidden_dim;
+        int num_heads;
+        int num_kv_heads;
+        int head_dim;
+        int max_seq_len;
         float_t rope_scaling;
         bool is_cache_enabled;
 };
